update ner algorithm
The ner_matching method now returns a dictionary with intents as keys and tuple[str, list[MatchedParam]] as values. For each intent, we have its ner_sentence (1st string of the tuple) and a list with the matched entities within the sentence (only those referenced in the intent are taken into account).

This way, we ensure that we obtain a set of matched parameters for each intent, which can be different depending on the intent since each intent can have a different set of referred entities.