update prediction
The prediction method now gets the new ner_matching_result. Now we have a set of different ner_sentences and for each ner_sentence a set of intents (the intersection of all the sets of intents is empty, and the union of all the sets intents is equal to the context intents).

For each ner_sentence, execute the prediction and assign the scores only to those intents associated to the ner_sentence.

Return a PredictResult object.