update base entity ner
base_entity_ner returns the formatted ner_sentence (which may be different from the original ner_sentence, e.g. "the level is minus three point four" ==> "the level is -3.4") and the formatted fragment (in the previous example: "-3.4"), so it can be later replaced properly by the entity name (in the previous example: "the level @SYS.NUMBER")

added support for negative numbers in ner_number function

fixed error in replace_param_value_with_entity_type_name_ner_sentence. When the fragment starts with '-' (it happens with negative numbers) the regular expression is different