update code structure
All text preprocessing methods are now within a dedicated file text_preprocessing.py. Some method names have been modified to be more generic

Tokenizers and stemmers are now created once (and not every time a text is preprocessed) and contained in a dedicated file pipelines.py

predict method preprocesses the sentence before the call to ner_matching (if applicable)

ner_matching looks for custom entities with preprocessed values (if applicable), so we can match variations of the original entity entries (e.g. match 'dogs' when the entry is 'dog')

CustomEntityEntry stores preprocessed values of 'value' and 'synonyms', since it is so expensive to compute this at every NER. This is computed in training.__train_context