update ner_matching algorithm
Intent.get_custom_entity_values_dict: a value/synonym can be referenced more than 1 time in an intent

ner_matching: we first replace found values by temporal strings, and then, in order of appearance, we replace them by the found entity and store the corresponding matched parameter. This way, we can know the order of appearance of parameters that refer to the same entity. An example:
'Are there more VALUE1 or VALUE2?' (both VALUE1 and VALUE2 refer to the same entity)
'Are there more men or women?': in ner_matching, we first would detect 'women', and then 'men' (because of the descendant alphabetical order of the values list), but we don't know which one is VALUE1 or VALUE2. Using the new approach, we would replace by 'Are there more /temp2/ or /temp1/?', and finally, we would replace temp values. While reading the sentence, we first find /temp2/, and since we know that the first entity reference in the intent is VALUE1, we create the 1st matched parameter, and then we match /temp1/ with the entity reference VALUE2. This approach considers the order of declaration of entity parameters in an intent.

add replace_temp_value_in_sentence and find_first_temp methods to utils.py